pyproject.toml
[project]
name = "spartdm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cvxpy>=1.7.1",
    "cvxpylayers>=0.1.9",
    "einops>=0.8.1",
    "h5py>=3.13.0",
    "ipykernel>=6.29.5",
    "jsonargparse[signatures]>=4.40.0",
    "kornia>=0.8.1",
    "leaspy==2.0.0rc2",
    "lightning[pytorch-extra]>=2.5.2",
    "lpips>=0.1.4",
    "matplotlib>=3.10.3",
    "monai>=1.3.2",
    "monai-generative>=0.2.3",
    "nibabel>=5.3.2",
    "omegaconf>=2.3.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "psutil>=7.0.0",
    "pyyaml>=6.0.2",
    "rich==14.0.0",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "tensorboard>=2.19.0",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "tqdm>=4.67.1",
]
